Bella: switch to Apache 2.0
The Apache license says “do whatever you want with this, just don’t sue me”
but does so with many more words :3, which lawyers like because it adds specificity.
It also contains a patent license and retaliation clause which is designed to prevent
patents (including patent trolls) from encumbering the software project.